requirements.txt
azure-storage-blob==2.1.0
Django==3.0.7
django-cors-headers==3.2.1
django-countries==6.1.2
django-filter==2.2.0
django-phonenumber-field==4.0.0 
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
djoser==2.0.3
djangorestframework-simplejwt==4.4.0
flake8==3.7.9
gunicorn==20.0.4
Markdown==3.1.1
phonenumberslite==8.12.7
Pillow==7.2.0
psycopg2==2.8.4
ptvsd==4.3.2
pytest==5.4.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3
requests==2.24.0